Issue 55.Change incr/dec method's signature,change parameter from int type to long type.

+++ src/main/java/net/rubyeye/xmemcached/CommandFactory.java
@@ -85,7 +85,7 @@ public interface CommandFactory {
 			CountDownLatch latch, CommandType cmdType, Transcoder<T> transcoder);
 
 	public abstract Command createIncrDecrCommand(final String key,
-			final byte[] keyBytes, final int amount, int initial, int expTime,
+			final byte[] keyBytes, final long amount, long initial, int expTime,
 			CommandType cmdType, boolean noreply);
 
 	@SuppressWarnings("unchecked")
