+++ src/main/java/net/rubyeye/xmemcached/MemcachedClient.java
@@ -702,7 +702,7 @@ public interface MemcachedClient {
 	 * @throws InterruptedException
 	 * @throws MemcachedException
 	 */
-	public abstract long incr(final String key, final int num)
+	public abstract long incr(final String key, final long num)
 			throws TimeoutException, InterruptedException, MemcachedException;
 
 	/**
@@ -721,7 +721,7 @@ public interface MemcachedClient {
 	 * @throws InterruptedException
 	 * @throws MemcachedException
 	 */
-	public abstract long decr(final String key, final int num)
+	public abstract long decr(final String key, final long num)
 			throws TimeoutException, InterruptedException, MemcachedException;
 
 	/**
@@ -966,7 +966,7 @@ public interface MemcachedClient {
 	 * @throws InterruptedException
 	 * @throws MemcachedException
 	 */
-	public void incrWithNoReply(final String key, final int num)
+	public void incrWithNoReply(final String key, final long num)
 			throws InterruptedException, MemcachedException;
 
 	/**
@@ -983,7 +983,7 @@ public interface MemcachedClient {
 	 * @throws InterruptedException
 	 * @throws MemcachedException
 	 */
-	public void decrWithNoReply(final String key, final int num)
+	public void decrWithNoReply(final String key, final long num)
 			throws InterruptedException, MemcachedException;
 
 	/**
